Implement AC-3.1 and improve inference complexity
This commit changes the implementation of `AC3` to an algorithm known as
AC-3.1 (Zhang, Yap). As a consequence the complexity of the arc
consistency drops to `O(n^2 * d^2)`, that is by a factor of `d`.

One drawback is an increased space required by the arc consistency. For
more info see the original paper.